Move EventLoopGroup.performWithTask(_:) to its own file, reimplement it in terms of the added EventLoop.performWithTask(_:), match documentation comments to those for the underlying EventLoopPromise.completeWithTask(_:), do some additional cleanup: - Rename FutureExtensions to Future+Miscellaneous for consistency - Move the `ELF.nonEmpty...()` extensions to their own file - Quick style fixup in Future.tryFlatMap()
