+++ Sources/AsyncKit/EventLoop/EventLoopGroup+Throwing.swift
@@ -26,20 +26,4 @@ extension EventLoopGroup {
     public func tryFuture<T>(_ work: @escaping () throws -> T) -> EventLoopFuture<T> {
         return self.next().submit(work)
     }
-    
-#if compiler(>=5.5) && canImport(_Concurrency)
-    /// Performs an async work and returns the result in form of an `EventLoopFuture`.
-    ///
-    /// - Parameter work: The async, potentially throwing closure to execute as a
-    ///   future. If the closure throws, a failed future is returned.
-    @available(macOS 12, iOS 15, watchOS 8, tvOS 15, *)
-    public func performWithTask<T>(
-        _ work: @escaping @Sendable () async throws -> T
-    ) -> EventLoopFuture<T> {
-        let promise = self.next().makePromise(of: T.self)
-        promise.completeWithTask(work)
-        return promise.futureResult
-    }
-#endif
 }
-
